Add residual debug output for batch equivalence testing results
- Implemented a new file `residual_debug.txt` to log the results of batch equivalence tests.
- The log includes details of each test case, including seeds, match results, and specific line discrepancies.
- Summary statistics for total tests, passed tests, failed tests, and execution time are also included.
- Preserved fixture files for further debugging in the specified temporary directory.

+++ ApogeeVGC/Sim/Actions/BeforeTurnAction.cs
@@ -7,7 +7,7 @@ public record BeforeTurnAction : IAction
 {
     public ActionId Choice => ActionId.BeforeTurn;
     public IntFalseUnion Order => 4; // Showdown: orders.beforeTurn = 4
-    public int Priority => 0;
+    public double Priority => 0;
     public int Speed => 0;
     public int SubOrder => 0;
     public int EffectOrder => 0;
