+++ ApogeeVGC/Sim/Actions/ResidualAction.cs
@@ -7,7 +7,7 @@ public record ResidualAction : IAction
 {
     public ActionId Choice => ActionId.Residual;
     public IntFalseUnion Order => 300; // Showdown: orders.residual = 300
-    public int Priority => 0;
+    public double Priority => 0;
     public int Speed => 0;
     public int SubOrder => 0;
     public int EffectOrder => 0;
